+++ R/ggMarginal.R
@@ -14,7 +14,8 @@
 #' provided and the \code{x} aesthetic is set in the main plot.
 #' @param y The name of the variable along the y axis. Optional if \code{p} is
 #' provided and the \code{y} aesthetic is set in the main plot.
-#' @param type What type of marginal plot to show. One of: [density, histogram, boxplot, violin].
+#' @param type What type of marginal plot to show. One of: [density, histogram, boxplot, violin, densigram] 
+#' (a "densigram" is when a density plot is overlaid on a histogram).
 #' @param margins Along which margins to show the plots. One of: [both, x, y].
 #' @param size Integer describing the relative size of the marginal plots
 #' compared to the main plot. A size of 5 means that the main plot is 5x wider
@@ -67,6 +68,9 @@
 #'
 #' # Using violin plot
 #' ggMarginal(p2, type = "violin")
+#' 
+#' # Using a "densigram" plot
+#' ggMarginal(p2, type = "densigram")
 #'
 #' # specifying the data directly instead of providing a plot
 #' ggMarginal(data = df, x = "x", y = "y")
@@ -99,7 +103,9 @@
 #'
 #' @seealso \href{http://daattali.com/shiny/ggExtra-ggMarginal-demo/}{Demo Shiny app}
 #' @export
-ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot", "violin"),
+ggMarginal <- function(p, data, x, y, 
+                       type = c("density", "histogram",  "boxplot", "violin", 
+                                "densigram"),
                        margins = c("both", "x", "y"), size = 5,
                        ..., xparams = list(), yparams = list(),
                        groupColour = FALSE, groupFill = FALSE) {